// packages/types/tests/schema-tests-3.2/comp-pathitems.ts
// Generated from https://github.com/OAI/OpenAPI-Specification/tree/v3.2-dev/tests/schema/pass/comp_pathitems.yaml
// Do not edit by hand; regenerate instead.
import type { OpenAPIObject } from '../../src/v3.2'

export const doc: OpenAPIObject = {
  openapi: '3.2.0',
  info: {
    title: 'API',
    version: '1.0.0',
  },
  components: {
    pathItems: {},
  },
}