// packages/types/tests/schema-tests-3.2/info-summary.ts
// Generated from https://github.com/OAI/OpenAPI-Specification/tree/v3.2-dev/tests/schema/pass/info_summary.yaml
// Do not edit by hand; regenerate instead.
import type { OpenAPIObject } from '../../src/v3.2'

export const doc = {
  openapi: '3.2.0',
  info: {
    title: 'API',
    summary: 'My lovely API',
    version: '1.0.0',
  },
  components: {},
} satisfies OpenAPIObject